+++ internal/consts/consts.go
@@ -373,6 +373,9 @@ const (
 	FieldPermanentlyDelete             = "permanently_delete"
 	FieldSignInAudience                = "sign_in_audience"
 	FieldTags                          = "tags"
+	FieldCustomTags                    = "custom_tags"
+	FieldSecretNameTemplate            = "secret_name_template"
+
 	/*
 		common environment variables
 	*/
@@ -445,6 +448,7 @@ const (
 	VaultVersion114 = "1.14.0"
 	VaultVersion115 = "1.15.0"
 	VaultVersion116 = "1.16.0"
+
 	/*
 		Vault auth methods
 	*/